cRUD for other college admnin

+++ Innovision/apis/CollegeList/Others/update.php
@@ -1,26 +1,29 @@
 <?php
 include('../../db.php');
-	$id = $_GET['id'];
-	$query = "SELECT college FROM other_college WHERE college_id=".$id;
-	$res = mysqli_query($conn, $query);
-
-	if(mysqli_query($conn, $query))
+	$id = $_POST['id'];
+	$new_name = $_POST['name'];
+	echo $new_name;
+	$query0 = "UPDATE other_college SET college = '".$new_name."' WHERE college_id=".$id;
+	if(mysqli_query($conn, $query0))
 	{
-		$arr = mysqli_fetch_array($res, MYSQLI_NUM);
+		$query = "SELECT college FROM other_college WHERE college_id=".$id;
+		$res = mysqli_query($conn, $query);
 
-		$query2 = "INSERT INTO college (college) VALUES ('".$arr[0]."')";
-		$res2 = mysqli_query($conn, $query2);
-
-		if(mysqli_query($conn, $query2))
+		if(mysqli_query($conn, $query))
 		{
-			$query3 = "DELETE FROM other_college WHERE college_id=".$id;
-			$res3 = mysqli_query($conn, $query3);
+			$arr = mysqli_fetch_array($res, MYSQLI_NUM);
 
-			if(mysqli_query($conn, $query3))
+			$query2 = "INSERT INTO college (college) VALUES ('".$arr[0]."')";
+			if(mysqli_query($conn, $query2))
 			{
-				echo(json_encode(array('status' => 'success', 'result' => 'Updated')));
+				$query3 = "DELETE FROM other_college WHERE college_id=".$id;
+
+				if(mysqli_query($conn, $query3))
+				{
+					echo(json_encode(array('status' => 'success', 'result' => 'Updated')));
+				}
 			}
 		}
-	}
+}
 
 ?>
